Change the allocation of "two-dimensional arrays". Rather than allocating an array of pointers, and then allocating memory to set each pointer, allocate all the memory at once for all the rows at once and compute the start of each row, rather than looking it up in an auxiliary table.  This is accomplished by the creation of a template Table<type> that remembers where the memory begins and also the number of bytes allocated to echo row.  The [] operator applied to a table compute the beginning of the appropriate row and returns it.
Most of this change is changing class member and function parameter
declarations from the old type to the new, always by reference
and not by value, so that this is no implicit copying of tables.
Where a reference can't be used - as when setting an element of
a class after the class object has already been constructed -
a pointer replaces the preferred reference member.


Former-commit-id: 6d84785cbd07e328e1f838de074fbb1201d65ca8