update comlier options to address a thread migration issue
With the old settings, the worker threads are migrating among cores after they are pinned to a core. This weird behavior is caused by Intel MKL library and is solved by updating the compiler options and linked libraries based on the MKL link guide. With the updated settings, there is no need to set the MKL_THREADING_LAYER to be sequential in the code.