Test: Add run script for testing MVP
This script `mvp_test.sh' will record Agora's output and dump to a
file under `log/' with a timestamp.

The script is used for debugging purpose since enabling DEBUG flag
in `cmake` will cause the system to crash in runtime. (Rx tiemout)
This might happen because the debug instructions add additional
time so the core operation cannot meet the deadline.

However, it seems the augmentation to UHD cause unexpected error in
the simulation mode. Therefore, we will have to test this UHD
version with RRU mode (real device).

MVP: minimum viable product.

Former-commit-id: b9b082989bd91ea0e2d2fe542600ecdba258af94