Fix: Manage ptr in safer way
Solution:
Using `make_shared` to new a pointer wil generate `shared_ptr`. It is a smart
pointer that manages its own lifecycle. In this case, using `get()` to obtain
the actual pointer is not recommended. Therefore, I updated it with `std::move()`
so that the ownership will be passed to the container vector correctly.

Bug:
Pointers in `computers_vec` become invalid in run func, while it was valid in
init func. Particularly, `DoDecode*` cannot be referenced.

Former-commit-id: bfebb2e1ec7a74c0e6776d0cde1edcd54d6980bd