Fix: Cali for BER/BLER
Bring back the built-in CFO calibration and align TX with RX signal to have
reasonable BER and BLER on RRU mode.

Note that we need 3 uplink symbols (U in fram_schedule in config) to have high-
quality built-in CFO calibration.

Also note that TX/RX alignment is device- specific and is measured with the
following steps:

1. Enable HDF5 recoder with cmake
2. Have a test run
3. Go to commit fb9ad884
4. Use ./files/log/plot_hdf5.py to read the latest log in
   `trace-uplink-YYYY-MM-DD-hh-mm-ss_1_1x1.hdf5` and set --ref-ul-slot=1
5. Read the top row of Fig.1 and see how far of the starting x is from 160.
6. Apply this difference to tx_advance in config file.


Former-commit-id: 91e842cd179e4c7ff2dc991cd40394896cfcf5ad